Create servlet for attaching a ta class for the user.
