+++ src/main/webapp/dashboard.html
@@ -40,7 +40,7 @@
       <div class="centered">
         <div class="ta-form">
           <!--Attach the TA email to the class list on submission-->
-          <form action="/add-ta" method="POST" onsubmit="getClassCode()" >
+          <form action="/add-class-ta" method="POST" onsubmit="getClassCode()" >
             <fieldset><legend>Add a TA by Email</legend>
               <label for="TAemail">
                 <span>Email <span class="required">*</span></span>
